refactor: remove re-export stubs and update imports (Phase 4e)
- Update all imports to point directly to new locations:
  - utils/collision → features/grid-editor/utils/collision
  - utils/fill → features/grid-editor/utils/fill
  - utils/validation → shared/utils/validation
  - Grid component imports → features/grid-editor/components/Grid/
  - Grid hook imports → features/grid-editor/hooks/

- Delete re-export stub files:
  - src/utils/collision.ts, fill.ts, validation.ts
  - src/hooks/useGridCoords.ts, useGridZoom.ts, etc. (9 files)
  - src/components/Grid/Bin.ts, GridCanvas.ts, etc. (4 files)

- Update test imports to point to new module locations

This completes Phase 4 of the folder structure reorganization.
