--- src/components/Grid/Bin.ts
@@ -1,2 +1,0 @@
-// Re-export for backward compatibility - supports direct imports and vi.mock paths
-export { Bin } from '../../features/grid-editor/components/Grid/Bin';
