--- src/components/Grid/GridCanvas.ts
@@ -1,2 +1,0 @@
-// Re-export for backward compatibility - supports direct imports and vi.mock paths
-export { GridCanvas } from '../../features/grid-editor/components/Grid/GridCanvas';
