--- src/components/Grid/IsometricPreview/Scene.ts
@@ -1,3 +1,0 @@
-// Re-export for backward compatibility - supports type imports from old paths
-export { Scene } from '../../../features/grid-editor/components/Grid/IsometricPreview/Scene';
-export type { SceneHandle, CameraPreset } from '../../../features/grid-editor/components/Grid/IsometricPreview/Scene';
