+++ src/components/Mobile/MobileLayersPanel/LayersTab.tsx
@@ -3,7 +3,7 @@ import { createPortal } from 'react-dom';
 import { useShallow } from 'zustand/shallow';
 import { useLayoutStore, useUIStore, useUndoableAction } from '../../../core/store';
 import { CONSTRAINTS, STAGING_ID } from '../../../core/constants';
-import { getDisplayLayers } from '../../../utils/collision';
+import { getDisplayLayers } from '../../../features/grid-editor/utils/collision';
 import { ConfirmDialog } from '../../../shared/components/ConfirmDialog';
 import { isOk, isErr, getUserMessage } from '../../../core/result';
