+++ src/core/storage/LayoutManager.ts
@@ -21,7 +21,7 @@
  */
 
 import * as backend from './backend';
-import { validateImport } from '../../utils/validation';
+import { validateImport } from '../../shared/utils/validation';
 import { generateLayoutId } from '../../shared/utils';
 import { STAGING_ID, CONSTRAINTS } from '../constants';
 import type {
