+++ src/core/store/view.ts
@@ -1,6 +1,6 @@
 import { create } from 'zustand';
 import { CONSTRAINTS } from '../constants';
-import { clamp } from '../../utils/validation';
+import { clamp } from '../../shared/utils/validation';
 
 /**
  * View Store
