+++ src/features/grid-editor/components/Grid/GridCanvas.tsx
@@ -4,7 +4,7 @@ import { useShallow } from 'zustand/shallow';
 import { useUIStore, useLayoutStore } from '../../../../core/store';
 import { useGridCoords, useGridTemplate } from '../../../../hooks';
 import { Bin } from './Bin';
-import { getBlockedZones } from '../../../../utils/collision';
+import { getBlockedZones } from '../../utils/collision';
 import { DEFAULT_CATEGORY_COLOR } from '../../../../core/constants';
 import type { Coord, ResizeHandle } from '../../../../core/types';
