+++ src/features/grid-editor/hooks/useGridCoords.ts
@@ -3,7 +3,7 @@ import type { RefObject } from 'react';
 import type { Coord } from '../../../core/types';
 import { useUIStore, useLayoutStore } from '../../../core/store';
 import { getBaseCellSize, snapToHalf } from '../../../core/constants';
-import { clamp } from '../../../utils/validation';
+import { clamp } from '../../../shared/utils/validation';
 import { useResponsive } from '../../../shared/hooks';
 
 /**
