+++ src/features/grid-editor/hooks/useGridResize.ts
@@ -3,7 +3,7 @@ import { useShallow } from 'zustand/shallow';
 import { useLayoutStore } from '../../../core/store';
 import { useUndoableAction } from '../../../core/store';
 import { CONSTRAINTS, STAGING_ID } from '../../../core/constants';
-import { clamp } from '../../../utils/validation';
+import { clamp } from '../../../shared/utils/validation';
 
 /**
  * Grid Resize Hook
