+++ src/hooks/interactions/useDrawInteraction.ts
@@ -1,6 +1,6 @@
 import { useCallback } from 'react';
 import { useLayoutStore, useInteractionStore, useHalfBinModeStore } from '../../core/store';
-import { canPlaceBin } from '../../utils/validation';
+import { canPlaceBin } from '../../shared/utils/validation';
 import { isOk } from '../../core/result';
 import type { InteractionContext, ModeHandlers, DrawStartArgs } from './types';
 import type { Coord } from '../../core/types';
