+++ src/hooks/interactions/useResizeInteraction.ts
@@ -1,6 +1,6 @@
 import { useCallback } from 'react';
 import { useInteractionStore, useHalfBinModeStore } from '../../core/store';
-import { canPlaceBin } from '../../utils/validation';
+import { canPlaceBin } from '../../shared/utils/validation';
 import { calculateResizeRect } from '../../utils/interaction';
 import type { InteractionContext, ModeHandlers, ResizeStartArgs } from './types';
 import type { Coord, Rect } from '../../core/types';
