--- src/hooks/useGridAxisLabels.ts
@@ -1,3 +1,0 @@
-// Re-export for backward compatibility
-export { useGridAxisLabels } from '../features/grid-editor/hooks/useGridAxisLabels';
-export type { GridAxisLabelsState, UseGridAxisLabelsOptions } from '../features/grid-editor/hooks/useGridAxisLabels';
