--- src/hooks/useGridCoords.ts
@@ -1,2 +1,0 @@
-// Re-export for backward compatibility
-export { useGridCoords } from '../features/grid-editor/hooks/useGridCoords';
