--- src/hooks/useGridFirstUseHints.ts
@@ -1,3 +1,0 @@
-// Re-export for backward compatibility
-export { useGridFirstUseHints } from '../features/grid-editor/hooks/useGridFirstUseHints';
-export type { GridFirstUseHintsState, UseGridFirstUseHintsOptions } from '../features/grid-editor/hooks/useGridFirstUseHints';
