--- src/hooks/useGridNavigation.ts
@@ -1,2 +1,0 @@
-// Re-export for backward compatibility
-export { useGridNavigation } from '../features/grid-editor/hooks/useGridNavigation';
