--- src/hooks/useGridResize.ts
@@ -1,3 +1,0 @@
-// Re-export for backward compatibility
-export { useGridResize } from '../features/grid-editor/hooks/useGridResize';
-export type { GridResizeState, UseGridResizeOptions, ResizeDirection, PendingResize } from '../features/grid-editor/hooks/useGridResize';
