--- src/hooks/useGridRowColumnSelection.ts
@@ -1,3 +1,0 @@
-// Re-export for backward compatibility
-export { useGridRowColumnSelection } from '../features/grid-editor/hooks/useGridRowColumnSelection';
-export type { GridRowColumnSelectionState, UseGridRowColumnSelectionOptions } from '../features/grid-editor/hooks/useGridRowColumnSelection';
