--- src/hooks/useGridTemplate.ts
@@ -1,3 +1,0 @@
-// Re-export for backward compatibility
-export { useGridTemplate } from '../features/grid-editor/hooks/useGridTemplate';
-export type { GridTemplateState, UseGridTemplateOptions } from '../features/grid-editor/hooks/useGridTemplate';
