--- src/hooks/useGridZoom.ts
@@ -1,3 +1,0 @@
-// Re-export for backward compatibility
-export { useGridZoom } from '../features/grid-editor/hooks/useGridZoom';
-export type { GridZoomState, UseGridZoomOptions } from '../features/grid-editor/hooks/useGridZoom';
