--- src/hooks/useInteraction.ts
@@ -1,2 +1,0 @@
-// Re-export for backward compatibility
-export { useInteraction } from '../features/grid-editor/hooks/useInteraction';
