+++ src/hooks/useKeyboardResize.ts
@@ -1,7 +1,7 @@
 import { useEffect, useCallback, useState } from 'react';
 import { useUIStore, useLayoutStore, useUndoableAction } from '../core/store';
 import { useMutations } from '../shared/contexts';
-import { canPlaceBin } from '../utils/validation';
+import { canPlaceBin } from '../shared/utils/validation';
 import { findBinById } from '../utils/entity';
 import { CONSTRAINTS, STAGING_ID, hasFractionalDimensions } from '../core/constants';
