+++ src/hooks/useLayoutRouting.ts
@@ -2,7 +2,7 @@ import { useEffect, useRef, useCallback } from 'react';
 import { useShallow } from 'zustand/shallow';
 import { useLayoutStore, useLibraryStore, useUIStore, useToastStore, useHistoryStore } from '../core/store';
 import { loadLayoutByIdAsync } from '../core/storage';
-import { validateLayoutIntegrity } from '../utils/validation';
+import { validateLayoutIntegrity } from '../shared/utils/validation';
 import {
   parseLayoutFromURL,
   setLayoutURL,
