+++ src/shared/hooks/useCrossTabSync.ts
@@ -1,7 +1,7 @@
 import { useEffect } from 'react';
 import { useLayoutStore, useLibraryStore, useHistoryStore, useUIStore, useLabsStore, LABS_STORAGE_KEY } from '../../core/store';
 import { loadLayoutByIdAsync, loadLibrary } from '../../core/storage';
-import { validateLayoutIntegrity } from '../../utils/validation';
+import { validateLayoutIntegrity } from '../utils/validation';
 import { createDefaultLabsPreferences } from '../../core/labs/types';
 
 /**
