+++ src/shared/hooks/useKeyboard.ts
@@ -1,11 +1,11 @@
 import { useEffect, useCallback } from 'react';
 import { useUIStore, useLayoutStore, useHistoryStore, useLibraryStore, useUndoableAction, useToastStore } from '../../core/store';
 import { useMutations } from '../contexts';
-import { canPlaceBin } from '../../utils/validation';
+import { canPlaceBin } from '../utils/validation';
 import { validateBinRotation } from '../../utils/binLocation';
 import { validateHalfBinModeToggle } from '../../utils/halfBinConstraints';
 import { SHORTCUTS, STAGING_ID, hasFractionalDimensions } from '../../core/constants';
-import { useGridNavigation } from '../../hooks/useGridNavigation';
+import { useGridNavigation } from '../../features/grid-editor/hooks/useGridNavigation';
 import { isOk } from '../../core/result';
 
 /**
