+++ src/test/hooks/useGridNavigation.test.ts
@@ -1,6 +1,6 @@
 import { describe, it, expect, vi, beforeEach, afterEach } from 'vitest';
 import { renderHook, act } from '@testing-library/react';
-import { useGridNavigation } from '../../hooks/useGridNavigation';
+import { useGridNavigation } from '../../features/grid-editor/hooks/useGridNavigation';
 import { useLayoutStore } from '../../core/store/layout';
 import { useUIStore } from '../../core/store/ui';
 import { resetAllStores } from '../testUtils';
