--- src/utils/collision.ts
@@ -1,12 +1,0 @@
-// Re-export for backward compatibility
-export {
-  getDisplayLayers,
-  getLayerZStart,
-  getBin3DRect,
-  footprintsOverlap,
-  verticalRangesOverlap,
-  binsCollide,
-  getBlockedZones,
-  isInBlockedZone,
-  checkLayerReorderCollisions,
-} from '../features/grid-editor/utils/collision';
