--- src/utils/fill.ts
@@ -1,2 +1,0 @@
-// Re-export for backward compatibility
-export { fillAllWithSize, fillGaps, getLayerCoverage } from '../features/grid-editor/utils/fill';
