--- src/utils/validation.ts
@@ -1,17 +1,0 @@
-// Re-export for backward compatibility
-export {
-  isValidDrawer,
-  isValidLayer,
-  isValidBin,
-  isValidCategory,
-  canPlaceBin,
-  canPlaceBinResult,
-  validateImport,
-  validateImportResult,
-  validateLayoutIntegrity,
-  validateLayoutIntegrityResult,
-  validateCustomProperties,
-  validateCustomPropertiesResult,
-  clamp,
-  truncate,
-} from '../shared/utils/validation';
